Allow only independents of same group to enter a vehicle
As independent groups are their own faction in A3W it is not good consequent that other indi's can easily hijack their vehicles.
Although it can be fun to hijack an enemy indi tank (yes, I'm guilty) it think vehicle restrictions should be enforced on a group level for indis.