Makes glitching into objects harder when exiting a vehicle
This code change checks the unit's current position after exiting the vehicle using BIS_fnc_findSafePos with a very small radius. If the position returned is <map center>, then the code tries to find a safe position from the vehicle within 5m. If found, it moves the player unit to the new location.